fix the problem with methods + tests
